fixed bugs where parameters where interpreted as part of the link
Fixed that the parameters could be interpreted as part of the site link, thereofre the sitemap would return not found instead of the correct site. This broke the handling of parameters, by delegating the analysing to the server functions this has been fixed. In addition to this the dependencies from Arduino's string class in the siteMap have been removed, allowing usage of the files on different infrastructures