Sepparated the conenction from the server
Refactored out the Connection details from the server class.  This allows the replacement of the connection if the server should not communicate with for example a hardware serial, added a timeout so that the server will return false if it cannot connect to the netwrok and that the user then can handle this.